Add utility to help execute tests in batches
This patch adds a helper script to allow users to execute a set of tests
in a single command.

The default behavior will be to execute all tests for all installed
bundles. Optionally, an user can set the 'BUNDLES' environment variable
to a space separated list of bundles to be tested.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>