[any-bundle] Lets unbreak 'ldconfig'
This test was basically destroyed.

The first basically masked the return code of 'ldconfig' therefore would
always return sucess (unless deleting the temp file happens to
file).

The second runs the same command twice, if ldconfig were to output an
error, the reminder of the test case is irrelevant since wont be
executed. Not only that, instead of calling ldconfig with '-X' was in
fact creating a cache file on current folder called '-X' :-/

There is no need to "count" (actually just to see if it is not empty)
the size of stderr output since ldconfig does the right and exits with
non-zero return code.

The problem the second tried to solve was only not captured by the
first test because the first was destroyed.

Signed-off-by: Murilo Belluzzo <murilo.belluzzo@intel.com>